🏗️🚧 Spliting channel related service into a file name channel.
