+++ frontend/src/components/Inputs/ChannelInviteDialog.tsx
@@ -10,7 +10,7 @@ import Switch from "@mui/material/Switch";
 import { ButtonChannelInvite } from "../Buttons/ButtonChannelInvite";
 import { Channel, IUserForLeaderboard } from "../../interfaces/users";
 
-import userService from "../../services/user";
+import channelService from "../../services/channel";
 import { useLoginContext } from "../../context/LoginContext";
 
 import Dialog from "@mui/material/Dialog";
@@ -74,21 +74,21 @@ export function ChannelInviteDialog({
   );
 
   React.useEffect(() => {
-    userService
+    channelService
       .getChannelInvitableFriends(loginContext.userLogin, channel.id)
       .then((friends: IUserForLeaderboard[]) => {
         setFriends(friends);
       });
 
     socket.on("update-channel-content", () => {
-      userService
+      channelService
         .getChannelInvitableFriends(loginContext.userLogin, channel.id)
         .then((friends: IUserForLeaderboard[]) => {
           setFriends(friends);
         });
     });
     socket.on("update-relations", () => {
-      userService
+      channelService
         .getChannelInvitableFriends(loginContext.userLogin, channel.id)
         .then((friends: IUserForLeaderboard[]) => {
           setFriends(friends);
@@ -119,11 +119,9 @@ export function ChannelInviteDialog({
   const handleInvite = () => {
     setOpen(false);
     selectedFriends.forEach((friend: IUserForLeaderboard) => {
-      userService
+      channelService
         .inviteToChannel(loginContext.userLogin, channel.id, friend.login42)
-        .then(
-					socket.emit("user:update-channel-content")
-				)
+        .then(socket.emit("user:update-channel-content"))
         .catch((err: Error) => {
           console.log(err);
         });
