+++ frontend/src/services/privateConv.tsx
@@ -2,11 +2,8 @@ import axios from "axios";
 import getConfig from "next/config";
 const { publicRuntimeConfig } = getConfig();
 const baseUrl = `http://${publicRuntimeConfig.HOST}:${publicRuntimeConfig.BACKEND_PORT}/privateConv`;
-import {
-  Message,
-} from "../interfaces/users";
+import { Message } from "../interfaces/users";
 axios.defaults.withCredentials = true;
-
 
 const createPrivateConv = (login: string, friendLogin42: string) => {
   const request = axios.post(`${baseUrl}/${login}/${friendLogin42}`);
